+++ Remember/ConjuntoMelhoriasFactory.java
@@ -1,8 +1,10 @@
 package Remember;
 
+import java.util.Set;
+
 public interface ConjuntoMelhoriasFactory {
 	
 	Componente setDuracao(Componente componente, String descricao, Double inicio, Double fim);
-	Componente setHabito(Componente componente, String descricao, Double inicio, Double fim);
+	Componente setHabito(Componente componente, String descricao, Double inicio, Double fim, Set<DiaDaSemana> diasDaSemana);
 
 }
