+++ Remember/DiaDaSemana.java
@@ -1,0 +1,11 @@
+package Remember;
+
+public enum DiaDaSemana {
+    DOMINGO,
+    SEGUNDA,
+    TERCA,
+    QUARTA,
+    QUINTA,
+    SEXTA,
+    SABADO;
+}
