+++ Remember/InterfaceFabrica.java
@@ -1,5 +1,9 @@
 package Remember;
 
+/*
+ * Interface para implementar o padrão Factory Method
+ * 
+ */
 public interface InterfaceFabrica {
     
     Componente getTarefa(int e);
