+++ Remember/Tarefa/ConjuntoMelhoriasTarefa.java
@@ -2,6 +2,8 @@ package Remember.Tarefa;
 
 import Remember.Componente;
 import Remember.ConjuntoMelhoriasFactory;
+import Remember.DiaDaSemana;
+import java.util.Set;
 
 public class ConjuntoMelhoriasTarefa implements ConjuntoMelhoriasFactory{
 	
@@ -11,8 +13,8 @@ public class ConjuntoMelhoriasTarefa implements ConjuntoMelhoriasFactory{
 	}
 	
 	@Override
-	public Componente setHabito(Componente componente, String descricao, Double inicio, Double fim) {
-		return new HabitoTafera(componente, descricao, inicio, fim);
+	public Componente setHabito(Componente componente, String descricao, Double inicio, Double fim, Set<DiaDaSemana> diasDaSemana) {
+		return new HabitoTarefa(componente, descricao, inicio, fim, diasDaSemana);
 	}
 
 }
